+++ Plugins/ChronicleEngine/Source/ChronicleEngineEditor/Public/Editor/ChronicleDialogueEditorLibrary.h
@@ -10,6 +10,42 @@ class UDialogueDatabase;
 class UDialogueRunner;
 
 USTRUCT(BlueprintType)
+struct CHRONICLEENGINEEDITOR_API FChronicleDialogueDebuggerVariableSnapshot
+{
+    GENERATED_BODY()
+
+    UPROPERTY(BlueprintReadOnly, Category="Chronicle|Debugger")
+    FGameplayTag VariableTag;
+
+    UPROPERTY(BlueprintReadOnly, Category="Chronicle|Debugger")
+    FVariableValue Value;
+
+    UPROPERTY(BlueprintReadOnly, Category="Chronicle|Debugger")
+    EChronicleVariableScope Scope = EChronicleVariableScope::Global;
+};
+
+USTRUCT(BlueprintType)
+struct CHRONICLEENGINEEDITOR_API FChronicleDialogueDebuggerEdgeSnapshot
+{
+    GENERATED_BODY()
+
+    UPROPERTY(BlueprintReadOnly, Category="Chronicle|Debugger")
+    FGuid TargetNodeGuid;
+
+    UPROPERTY(BlueprintReadOnly, Category="Chronicle|Debugger")
+    int32 FromSlotIndex = 0;
+
+    UPROPERTY(BlueprintReadOnly, Category="Chronicle|Debugger")
+    FString ConditionExpression;
+
+    UPROPERTY(BlueprintReadOnly, Category="Chronicle|Debugger")
+    bool bConditionPasses = true;
+
+    UPROPERTY(BlueprintReadOnly, Category="Chronicle|Debugger")
+    float Weight = 1.0f;
+};
+
+USTRUCT(BlueprintType)
 struct CHRONICLEENGINEEDITOR_API FChronicleDialogueDebuggerSnapshot
 {
     GENERATED_BODY()
@@ -24,6 +60,12 @@ struct CHRONICLEENGINEEDITOR_API FChronicleDialogueDebuggerSnapshot
     FGuid CurrentNodeGuid;
 
     UPROPERTY(BlueprintReadOnly, Category="Chronicle|Debugger")
+    EDialogueNodeType CurrentNodeType = EDialogueNodeType::Speech;
+
+    UPROPERTY(BlueprintReadOnly, Category="Chronicle|Debugger")
+    int32 CurrentLineIndex = INDEX_NONE;
+
+    UPROPERTY(BlueprintReadOnly, Category="Chronicle|Debugger")
     EDialogueRunnerState RunnerState = EDialogueRunnerState::Idle;
 
     UPROPERTY(BlueprintReadOnly, Category="Chronicle|Debugger")
@@ -31,6 +73,54 @@ struct CHRONICLEENGINEEDITOR_API FChronicleDialogueDebuggerSnapshot
 
     UPROPERTY(BlueprintReadOnly, Category="Chronicle|Debugger")
     FText NodeTitle;
+
+    UPROPERTY(BlueprintReadOnly, Category="Chronicle|Debugger")
+    TArray<FChronicleDialogueDebuggerVariableSnapshot> Variables;
+
+    UPROPERTY(BlueprintReadOnly, Category="Chronicle|Debugger")
+    TArray<FChronicleDialogueDebuggerEdgeSnapshot> OutgoingEdges;
+
+    UPROPERTY(BlueprintReadOnly, Category="Chronicle|Debugger")
+    TArray<FDialogueHistoryEntry> History;
+
+    UPROPERTY(BlueprintReadOnly, Category="Chronicle|Debugger")
+    TArray<FString> SeenDialogueHashes;
+};
+
+USTRUCT(BlueprintType)
+struct CHRONICLEENGINEEDITOR_API FChronicleConditionExpressionValidationResult
+{
+    GENERATED_BODY()
+
+    UPROPERTY(BlueprintReadOnly, Category="Chronicle|Expression")
+    bool bParsed = false;
+
+    UPROPERTY(BlueprintReadOnly, Category="Chronicle|Expression")
+    bool bEvaluationResult = false;
+
+    UPROPERTY(BlueprintReadOnly, Category="Chronicle|Expression")
+    TArray<FString> VariableReferences;
+
+    UPROPERTY(BlueprintReadOnly, Category="Chronicle|Expression")
+    FString Message;
+};
+
+USTRUCT(BlueprintType)
+struct CHRONICLEENGINEEDITOR_API FChronicleDialoguePipelineExportPaths
+{
+    GENERATED_BODY()
+
+    UPROPERTY(BlueprintReadOnly, Category="Chronicle|Pipeline")
+    FString JsonFilePath;
+
+    UPROPERTY(BlueprintReadOnly, Category="Chronicle|Pipeline")
+    FString LinesCsvFilePath;
+
+    UPROPERTY(BlueprintReadOnly, Category="Chronicle|Pipeline")
+    FString LocalizationCsvFilePath;
+
+    UPROPERTY(BlueprintReadOnly, Category="Chronicle|Pipeline")
+    FString AuditJsonFilePath;
 };
 
 UCLASS()
@@ -71,6 +161,15 @@ public:
 
     UFUNCTION(BlueprintPure, Category="Chronicle|Editor|Dialogue Tree")
     static FText GetNodeTypeDisplayName(EDialogueNodeType NodeType);
+
+    UFUNCTION(BlueprintCallable, Category="Chronicle|Editor|Expression")
+    static bool ValidateConditionExpressionForTree(UDialogueTree* Tree, const FString& Expression, FChronicleConditionExpressionValidationResult& OutResult, FString& OutError);
+
+    UFUNCTION(BlueprintCallable, Category="Chronicle|Editor|Pipeline")
+    static bool BuildDefaultDialogueTreeExportPaths(UDialogueTree* Tree, const FString& ExportDirectory, FChronicleDialoguePipelineExportPaths& OutPaths, FString& OutError);
+
+    UFUNCTION(BlueprintCallable, Category="Chronicle|Editor|Pipeline")
+    static bool ExportDialogueTreePipelineArtifacts(UDialogueTree* Tree, const FString& ExportDirectory, const FString& Culture, FChronicleDialoguePipelineExportPaths& OutPaths, FString& OutError);
 
     UFUNCTION(BlueprintCallable, Category="Chronicle|Editor|Debugger")
     static bool CaptureDebuggerSnapshot(UDialogueRunner* Runner, FChronicleDialogueDebuggerSnapshot& OutSnapshot, FString& OutError);